requirements.txt
numpy==1.24.2
pillow==9.4.0
matplotlib==3.6.3
tkinterdnd2==0.3.0
renishawWiRE==0.1.16
